Вынесли блог в отдельное приложение
