+++ catalog/templates/base.html
@@ -70,7 +70,7 @@
                 </div>
             {% endif %}
             <a class="p-2 text-dark" href="#">Сотрудничество</a>
-            <a class="p-2 text-dark" href="/posts">Блог</a>
+            <a class="p-2 text-dark" href="{% url 'blog:list_of_posts' %}">Блог</a>
             <a class="p-2 text-dark" href="/contacts/">Контакты</a>
             {% if user.is_authenticated %}
                 <a class="p-2 text-dark" href="{% url 'users:profile' %}">Профиль</a>
